Round 56 closes: four more coherence repairs, round document, receipt, tooling
5,958 pins in 1,264 rows; 11 recipe deviations repaid under charter
clause 4b; lanes ledger rebuilt; land_recipe_switch.sh for the
byte-neutral recipe reduction (round 56b).

+++ src/dungeon/func_800D5A04.c
@@ -30,7 +30,7 @@ extern s32 D_800814A0;
 
 /* Updates a timed effect's animation and motion, marking it for removal when finished. */
 void func_800DB164(void *effect, void *motion, void *visual) {
-    s16 phase;
+    s32 phase;
     s32 ticks_left;
     s32 next_phase;
